Refactor MiniscopeImagingExtractor to use static methods for file validation and configuration loading
